+++ functorch/compile/csrc/CompileCache.cpp
@@ -10,7 +10,7 @@
 /// allowing different types of hashing functions, and is agnostic of the
 /// compiler.
 ///
-#include <functorch/csrc/CompileCache.h>
+#include <functorch/compile/csrc/CompileCache.h>
 #include <torch/csrc/autograd/custom_function.h>
 #include <torch/csrc/jit/python/pybind_utils.h>
 #include <torch/csrc/jit/tensorexpr/codegen.h>
